+++ CHANGES.rst
@@ -9,6 +9,10 @@ astropy now requires a Numpy 1.6.0 or later.
 New Features
 ^^^^^^^^^^^^
 
+- ``astropy.analytic_functions``
+
+  - New module to store analytic functions [#3077].
+
 - ``astropy.config``
 
 - ``astropy.constants``